Fix Adafruit_GFX constructor-related link issue
Declare Adafruit_GFX::drawPixel pure virtual so that the compiler
knows not to expect a definition of the method in this class. The
compiler appears to create the vtable with the first
non-constructor, non-pure virtual method so when this method is
the first non-constructor method in the class and it is not
declared pure virtual the linker fails to find the vtable for this
class.

Change the Adafruit_GFX::constructor method to a real constructor.