Custom Fonts
An interim solution until a true API is released.  Allows for alternate
fonts and font sizes  Fonts are restricted to being a fixed width for
all characters.  Use the Google GLCDFontCreator tool and strip all the
metadata before passing in the font.

Hint: You must modify the width of every character in the tool
Hint: The constant FONTNAME_WIDTH is not precise if you imported a font.
Hint: The first character specifies the first character to which the
data is supplied.  For example, by default the tool exports only
starting at character with ASCII code 32.