Fix compile errors USE_FAST_PINIO + compile on Robotis boards
There were a couple of compile issus when you are trying to use a different board which requires you not to use fast pinio.  In particular Robotis OpenCM9.04 as well as the OpenCR 1.0 boards.

They are STM32 based __arm__ boards but do not have the support for setting/clearing IO pins using memory mapped access to one location.

First thing I ran into was the code in the #else casue was using
#undefine
which is not valid, should be #undef

Second issue is once this define was undefined, the #defines for SPI_CS_HIGH
and SPI_CS_LOW did not compile as there was no ; at the end of the digitalWrite.
Probably worked earlier until somone put this statement into a {} pair