fontconvert: remap characters, for smaller custom fonts
Add support for a new (optional) "remap" parameter for `fontconvert`.

If given, that parameter replaces the glyphs for certain characters
with glyphs of a different character.

This way a font can be created which contains a single contiguous
range of characters, without being limited to 7 bit ASCII.

The downside of this approach is, that the characters as they appear
in the code won't show up on screen the same way.

Example:

Suppose I need the following set of characters: `0123456789+-.%°`.
The first 14 characters are all in the ASCII plane, so without
remapping, I could get them by converting the range 37 (%) to 57 (9):

    ./fontconvert FreeSerifBold.ttf 24 37 57

(`°` is not on the ASCII plane, so I cannot get that character, no
matter the range).

Note that even though I needed only 15 characters, I had to specify a
20 character range to get only 14 of them. There are 6
characters (`&'()*,/`) for which glyphs will be included as well, but
I'll never use them.

This isn't ideal, so "remap" to the rescue!

First I'm choosing an ASCII range that includes as many of the
characters as I need, by having as little extra characters as
possible.
In this case that range is 43 (`+`) to 57 (`9`), which has two
characters I don't need: `,` and `/`. Luckily I also have two
characters which are *not* in that range (`%` and `°`).

By using the `°` glyph for the `,` character and the `%` glyph for the
`/` character, I get a contigious range of characters which contains
only the symbols I need:

| Charcode | ASCII | Glyph |
|----------|-------|-------|
| 43       | `+`   | `+`   |
| 44       | `,`   | `°`   |
| 45       | `-`   | `-`   |
| 46       | `.`   | `.`   |
| 47       | `/`   | `%`   |
| 48       | `0`   | `0`   |
| 49       | `1`   | `1`   |
| 50       | `2`   | `2`   |
| 51       | `3`   | `3`   |
| 52       | `4`   | `4`   |
| 53       | `5`   | `5`   |
| 54       | `6`   | `6`   |
| 55       | `7`   | `7`   |
| 56       | `8`   | `8`   |
| 57       | `9`   | `9`   |

The remap parameter for this case is `,°/%`:

    ./fontconvert FreeSerifBold.ttf 24 43 57 ,°/%

From arduino code then, I can just write:

  display.display("37.5,");

Which will render "37.5°".